Update UI styles and enhance README documentation
- Changed footer background color and text styles for better visibility.
- Updated button styles across various components for consistency and improved user experience.
- Added a requirement note for DATABASE_URL in the replit.md file.
- Minor formatting adjustments in README.md for clarity.
