+++ README.md
@@ -286,3 +286,4 @@ For support and questions:
 ---
 
 **Built with ❤️ for modern e-commerce**
+
