+++ client/src/pages/admin-fixed.tsx
@@ -791,3 +791,4 @@ const AdminPage = () => {
 };
 
 export default AdminPage;
+
